requirements.txt
fastapi==0.95.1
locust==2.16.1
loguru==0.7.0
numpy==1.24.3
pandas>=1.4.4
pydantic==1.10.7
pytest==7.4.0
python-multipart==0.0.6
PyYAML==6.0
scikit-learn==1.1.2
uvicorn==0.22.0
mlflow==2.3.2
xgboost==2.1.4
# trace package 
opentelemetry-api==1.19.0
opentelemetry-sdk==1.19.0
opentelemetry-instrumentation-asgi==0.40b0
opentelemetry-instrumentation-fastapi==0.40b0
opentelemetry-instrumentation-requests==0.40b0
opentelemetry-instrumentation-logging==0.40b0
opentelemetry-exporter-jaeger==1.19.0
opentelemetry-exporter-otlp-proto-grpc==1.19.0
opentelemetry-exporter-prometheus==1.12.0rc1
# prometheus client
prometheus-client==0.17.1
# Ansible
ansible==8.3.0
requests==2.31.0
google-auth==2.28.1
# Datalake-minio
fastparquet==2024.5.0
minio==7.1.16
PyYAML
# cdc-kafka
dotenv
psycopg2-binary
sqlalchemy
# spark
pyspark==3.3.2
pydeequ==1.0.1
python-dotenv==1.0.0
nltk==3.8.1
# Flink 
apache-flink==1.17.1
kafka-python==2.0.2
avro-python3==1.8.1
#pandas==1.3.0
pyarrow==8.0.0
python-schema-registry-client==2.4.1
pymongo==3.8.0
# Onnx 
onnx
onnxmltools==1.14.0
onnxruntime==1.24.4
# triton 
tritonclient==1.24.4
# kserve
kserve==0.15.0